_filter_csv_fids:
Fixes adding missing masters to the BP. I moved the updating of the
parser data structure to fit the patcher in the patcher - this is the
right thing but comes at the expense of a RecordType import. Note that
before the same lfid *could* be added multiple times with different
ranks to the list of factions in _update_from_csv. I factored out
_PatchFidReplacer from ReplaceFormIDsPatcher to do the _filter_csv_fids
check - this parser still completely escapes me - incidentally it was
this patcher that added the missing fids - from the bundled
P1DCandles_Formids.csv.