Make wizard and FOMOD images work on Linux
Namely by introducing a function to convert a case insensitive path to
the right case for the current filesystem. This is basically what Wine
does as well, but my guess is Wine caches a whole bunch to make this
more efficient. Not a big deal right now since we only use it for fixing
the image paths, which were designed with Windows in mind, but if we
want to use this to fix BAIN, we'll have to think of a more performant
solution.

Under #243