[!]Wip refactor _AParser:
Well a method called write_record_2 asks to be removed. This adds a new
method and class variable but sure looks cleaner (plus getDefault needs
itself encapsulation). As a bonus we stopped copying the cls_rel_attrs
on every row write.

The important change here however is in ActorFactions _update_from_csv.
When not called from the patcher it would only keep for an actor id
the last faction rank - this is certainly not what was intended but
adding an [!] just in case. See also bcef035048c0f8e09359bd9cf6e53bb7ffdb96cd
and links there (partly the csv work in this branch was motivated from
an attempt to restore csv support for ImportRelations which made me
look at the parsers again).