bain: Don't redirect docs in no_skip{,_regexes}
Doc-like files that shouldn't be skipped shouldn't be redirected for the
same reasons. This only affects files in Data (or Data Files), so
no_skip_dirs would defeat the purpose of the redirect to some extent.

FNV: Add aaid.txt to no_skip

Auto Activate Invisible Doors:
https://www.nexusmods.com/newvegas/mods/80650

Accessed by a script in the mod.