Escape HTML chars when converting wtxt
The GTK webview refuses to load further if e.g. an unescaped '&' exists
in the document. Firefox, Edge, etc. are forgiving there, but this *is*
obviously a mistake on our part - we should escape such characters to
'&amp;' etc.

requirements.txt:
Bump dependencies

Checked the changelogs, pretty much nothing but bug fixes and feature
additions that don't concern us.