Fix for screenshots renames:
The regex for screenshots compiled in ListInfo._name_re classmethod is:

    '^(?=.+\\.)(.*?)(\\d*)(\\.(?:tif|jpg|tga|gif|bmp|jpeg|png))$'

so it seems we only accept digits right before the extension. Probably
we wanted to allow for digits in the middle? Anyway this fixes adding an
extra extension - all this is probably due to the screenshots taken in
game that have a Screenshot[digits].ext IIRC.

Drop the f'{label}' call in PopulateItem - not needed anymore?

I believe this was a nightly regression in 543 development where the
installer label was stored as Path and not needed anymore but we would
need to update from a 310 install to check.

Gui 'label' renames -> header.label is what we are after here - thanks
Inf for fixups!

Co-authored-by: Infernio <infernio@icloud.com>