Load _ExteriorCells block by block:
This amends the correct but super hacky code in
29ad38ec6b00158054f6b0be04d1edb5aa9f7754. We trust the endBlockPos
instead of the end_pos of the world children group, effectively loading
block by block. We need then to allow for duplicate type 4 groups. By
the time the block is exhausted we have returned to the _Nested
_load_rec_group - ROAD and (persistent) CELL should be handled there.
This took down the isinstance hacks but also the grup_head_or_end (yak),
and the kwargs.get('endPos') (and a load_mobs override). These were the
ugliest remaining hacks in the _AMobBase hierarchy which enters a
proud beta (ok there is still _process_rec).

Under #647.

Mopy/bash/brec/mod_io.py: the only real difference turns out to be the
label. We avoid the append call and passing __rh, and we drop another
override. As a bonus pack_formats and header attrs are centralized -
as noted "The way we represent form versions in memory needs rethinking"

Under #480.