Fix Enter when renaming screenshots on Linux
Would always finish the rename operation *and* open the selected
screenshot. The reason is because we used _handle_key_up instead of
_handle_key_down, which MSW seems to tolerate while GTK really doesn't.
This was clearly wrong, seeing as all the other tabs with 'Enter to
open' use _handle_key_down as well.

Mopy/bash/gui/popups.py:
Fix fallback continue dialog sizing

That Stretch is not needed and will in fact break the dialog.

Under #243