_AMerger: Take deltas from masters into account
Consider this situation:
 - A.esp has Invent.Add, adds an item to a CONT record
 - B.esp touches the same CONT record, what it does is irrelevant
 - C.esp has Invent.Remove and has A.esp and B.esp as masters, removes
   the item A.esp added again

Since we only considered rule of one for calculating the delta, the
patcher saw the version from B.esp. But we'll end up merging in the
addition from A.esp as well, so we have to take that into account when
calculating the delta.

Had to break the delta-applying step out of buildPatch, which makes it
more readable anyways (and will hopefully help with rewriting it later -
see #312).