Add support for GOG versions of TES3, FO3 and FNV
Can't use game_detect_includes here because that one is a logical
conjunction. For the .ico files used by GOG games we need a logical
disjunction instead, so we have to override test_game_path. Also made
the mixins inherit from GameInfo to fix some PyCharm warnings.

Unsure if we can add a separate installer option for these. Someone
would have to test if they overwrite Steam's registry key (we'll
eventually want to refactor the way we detect Steam games to stop
relying on that potentially shared key, which will fix this problem for
good).

Under #646