Don't ghost the main game ESM
We rely on that one being present on boot. Could refactor WB to allow
the main game ESM to be ghosted, but I don't think allowing this one to
be ghosted is a good idea anyway - how is the game going to boot without
it? Even the original Nehrim needed an empty, dummy Oblivion.esm file to
be there.