Add support for listing default tweak 'errors'
More accurate terminology would be "reasons why the INI can't be
applied", but that's a different can of worms. The only reason why this
had to be disabled for default INIs is the usage of self.abs_path, but
we only use that to the get the tail - which we already have as
self.fn_key. So just use that, and it magically works for default tweaks
as well.