Rebuke the lord of undead (functions) getGhosted (and fixups):
Using the power of Store artifact (and the observation that if the file
exists then the store knows all about it including the path - note that
limbo ghosts, that have both normal and ghost versions remain an issue.
I changed return type of data_path_to_key -> info to use that info).
A break was missing in:

-                for store, removed_files in removed_tracked:
-                    if fname_key := store.data_path_to_key(path):
-                        removed_files.add(fname_key)
                         break # <----- here
-                else:
                     removed_untracked.add(path)
                     empty_dir_candidates.add(path.head)

removed_untracked.add(path) would always run (and the data store dirs
would be added to empty_dir_candidates) which looks like a bug. Note
also:

@@ -2730,3 +2730,3 @@ def _removeFiles(self, ci_removes, refresh_ui, progress=None):
             for store, removed_files in removed_tracked:
-                refresh_ui[store.unique_store_key] = True
+                refresh_ui[store.unique_store_key] = bool(removed_files)
                 store.delete(removed_files, recycle=False)

which will hopefully soon change to use a RefrData instance :)

__installer_install refactor mtime checks

We can't avoid them really but we can use the created info that is in
scope - TTT very subtle interplay between order of operations and order
of stores.

Correct data_path_to_info to not return True for data paths like
'subdir/some.[esp|bsa]' (if we get those). Based on the observation
that self.get() won't return if data_path is not a basename - note for
screens override that:

*a, b = [1] -> a=[], b=1

Adds a couple lower() but keeps the basename call for 'would_be' path.

Typing here needs more work but this is natural as it's the boundary
between FName and CIstr clients - time to standardize (including
ci_* prefixing) but first refresh.

PS: That was a heroic battle, seen in the positive line count -
getGhosted proved very hard to uproot and first draft of this was
probably *the* most broken commit in 312. Thanks Inf for pointing out
the bug - the solution was to transfer the processing in _fs_install,
which made __installer_install the thin wrapper it was supposed to be.

Under #219, at least - follows c6ec3997e3ef26ac350c47933567602ffdfc02e3.

Infernio:

- We need the actual store here, not the key
- data_tracking_stores was a generator - make it a tuple so using it is
less error prone.

wrye_text: Restore double underscore removal

Originally added in 68a218186a40712cc69a84cc613127c65d03f9b5, got
lost during the split in 2c9e69a55270f60f040b2ff5cff155b5ac6b7623.

Co-authored-by: Infernio <infernio@icloud.com>