Empty dirs handling:
Changes the logic:

- remove subdirs that contain no files in any of their subfolders - the
root one should be taken care of at the caller's level. Note the
(hacky) 'proj_dir.makedirs()' is not needed anymore, by simply not
handling the return value (compare this with handling it in has_files
for top Data dirs). Previously we would leave behind empty subfolders,
except if we carefully sorted, which might have been the case (or not).
Plus we could delete folders before their subfolders (hence removedirs
was called defensively and abundantly). Now the logic is clearly
spelled out in _remove_empty_dirs (deserves a test certainly) and
repeated in _walk_data_dir (as noted we need performance so factoring a
walk function parameter out is a no-no - plus we can afford some
repetition for the readability, especially as the walk functions are
inlined closer to irefresh and not buried as before).

- we remove as we go. This is less atomic but thankfully we had no
guaranties anyway :Plus we won't really miss anything - if the operation
fails we just leave behind less empty dirs.

Kept error handling the same as before - although all the dirs should
exist in the `raise_error=False` case in _walk_data_dir.