FO4: Implement NPC_
Includes a ton of renames related to NPC attributes. Also came with a
fun surprise in the form of MelActorSounds2. Turns out xEdit doesn't
properly decode them, which is not terribly surprising because Bethesda
chose a really weird way to implement them. See _MelCs2kCs2d for more
info.

With that, one of the nastiest records is done.