Fix Import Face on a plugin
Several issues:
 - We were putting tons of short FormIDs into the ModFile, which blew
   up when it tried to write those. To that end, I made MasterMap spit
   out only FormId instances but accept both FormIds and ints to make
   transitioning faces.py and _saves.py over to only FormIds easier (see
   below).
 - We were using from_object_id to create a new record, but that created
   one with only a short FormID (duh). Plus it was completely redundant,
   create_record already takes care of making a proper new FormId for
   us.
 - We only set 'health' and 'unused2' values when the face had truthy
   'health' information, relying on defaults otherwise. Those defaults
   are not set anymore, of course.

This indicates that more of faces.py and _saves.py needs to transition
to using FormId instances, but that's not exactly news. Probably breaks
other places that were using MasterMap (saves_links.py, most notably),
though I did try to go through the logic in those places (but haven't
tested it).

We are now down to a single from_object_id usage, by the way. We should
see if we can't obsolete that.