FO4: Port Import Actors: Perks
Let's start tackling this monster.

Isn't it hilarious that this commit ports a patcher to a new game and
has a negative line diff? Granted, that's only because of the docs
changes, but still.

Also change the default for spell_types to include LVSP. I excluded that
because FO4 didn't have it in xEdit's definitions back then and I wanted
the defaults to be futureproof, but it turns out FO4 does have LVSP,
it's just unused. So we need to specify it anyways, which means only FO3
does not have LVSP.