Fix MelExtra causing corruption
It wrote out the extra data but did not bump up the subrecord size. Must
have broken during some refactoring. No one noticed because it was only
used for NAVI and DOBJ - we never write the former and I doubt we ever
write the latter either. FO4 does use it for something common though,
namely NPC_.

The fix is slightly complex because we have to bypass
_MelWrapper.dumpData and redirect _MelWrapper.packSub to the wrapped
Mel*, so that we can intercept pack_subrecord_data.