Don't freeze on encoding detection
If we're ever fed a truly huge bytestring (used >16KB here, which
should be far beyond any string we ever sanely feed in here), then
chardet can cause the entirety of WB to hang. Avoid that by feeding it
only some data at a time and aborting as soon as we get a usable result.
This may change the detection, but we never feed in such huge
bytestrings unless we're given malformed input data (in this case, an
INI file that contined 7MB of binary junk).

Mopy/bash/bosh/__init__.py:
Fix crash when LOOT is not installed on Linux/Mac

Forgot to set a tooldir for that case.

Mopy/bash/belt.py:
Don't traceback when hitting wizard parse error

Otherwise we get a second window containing the same message and a
traceback, which looks like WB hit an unexpected error of some kind
(when it's actually the wizard script that's broken).