Remove fixedFont param:
The txtCtrl.SetStyle was commented out - only set to True in
Mopy.bash.basher.mod_links.Mod_Details - test if it made a difference
Setting Fonts is priority (very) near 0 but if we ever do (don't) we
should have some FontOptions class to pass around - could be done with
Style stuff too maybe.

This and next are in preparation for the _Log-to-gui commit.

Under #173