WIP Unify handling of pos/size_key:
The few uses would always append .pos and .size to the same _key_prefix,
so standardize that.

Align kwargs of Frame/Dialog with _topLevelWin

Changed the key to end with .size:

@@ -161,4 +161,4 @@ def __init__(self, parent=None, title=None, *, sizes_dict=None,
                  stay_over_parent=False, **kwargs):
-        self._size_key = f'{self._key_prefix}.size' if self._key_prefix else \
-            self.__class__.__name__
+        sk = self._key_prefix if self._key_prefix else self.__class__.__name__
+        self._size_key = f'{sk}.size'
         self._pos_key = self._key_prefix and f'{self._key_prefix}.pos'

This would reset some sizes (if saved in settings) but not worth the
backwards compat. Should we also save pos always (but see TODO)?
Maybe create a "pos" dict also for easy resetting?

TODO:
- add a setting for reseting pos and size for all windows. This will help
when a dialog appears outside the screen which effectively freezes it
for the user - they would need to delete BashSettings.dat/bak in this
case.