Move _Log to gui:
Since we got rid of sizes this became possible. This is the final
result, I tried various approaches as it was still tricky (note the
differing default params for instance).  I decided to keep the
_show[Wrye]Log API but made it a thin wrapper. We should address the
todo's in _LogWin.__init__.

_show[Wrye]Log were never passing an lg_icons in (apart from a direct
use of WryeLog that was passing bashBlue) so I removed the parameter
Note that `_key_prefix` needs now be an instance attribute.