Refactor BAIN refresh, prepare for Store
A lot of BAIN refactoring here to un-hardcode the mods/inis assumptions.
These are already violated by the BSAs tracking ownership as well.

Note the really ugly _is_ini_tweak magically disappearing by itself due
to this.

APIs are still pretty rough here, rest of the branch will loosen the
coupling, formalize the whole thing, etc.

Screenshots can be Data-relative

In the interest of getting this right, let's get every eventual
consumer of these APIs in the room.

Under #219 and #353