Fixup for possible bug in EventHandler:
Mopy/bash/gui/events.py:
The docs claim that """The order in which the listeners are processed is
guaranteed to be the same as the order in which they subscribed""" but
pause_subscription would always append possibly violating this condition.

Rest is nit and some fdust - I also changed the docstring interpreter
part as it would not resolve the import - 'images' renames and a stray
bass import (only one local bass import in gui, images related)

Nit and renames from the future - especially from the Store(Enum) work:

Chop off one use of tabInfo and extract some local variables to more
easily track the usesites

bosh: typing nit in _parse_save_path

Note I dropped the bsa.refresh on creating the BSAPanel - we do that in
InitData and RefreshData (and anyway it's disabled). I kinda tried
nuking the BashTab inits (most exist just to set listData) but did
not come up with something satisfactory (remember this is from future,
the BashFrame.***List = self.uiList go bye-bye).

env: some nit split out from the ask_confirm commit for readability