Decouple bosh from balt:
This does it finally (at the expense of dropping the progress from the
refresh call - note this would pop up on tab in for instance). Reduced
duplication around rescanMergeable.

Drop balt from converters:

refreshConverters was binding us - single use in irefresh. I pass a
progress in there. A very few progresses may disappear - oh well, plus
they would just flash by as converters are usually few to none, plus
once this is frozen we can revisit progress handling.
Edit:the only call sites of irefresh this could make a difference IIUC:

- refresh_n/refresh_ns are  - but then should we run refreshConverters
at all? This is the FIXME in irefresh itself, we should revisit the
`if changed or ...` and be more specific.
- the calls of irefresh(what='I') in bain - this should be handled
someplace central and eventually be passed a Progress from the UI.

Drop balt from mods_metadata:

This change is a bit more tricky as checkMods does a few things before
the progress was created and so a CancelError could be thrown outside
the try catch. Leaving separate to review

And with this and that ___bosh is decoupled from gui___

Under #190.