Rework event handling for status bar buttons DnD:
Remove the mouse capture code:

We need to test this on windows but miraculously works on __WXOSX__ and
maybe also in __WXGTK__? -> yes
What was happening is that somehow the event.Skip() in OnDragStart was
propagating the event which somehow blocked the OnDrag from being called

Removing the event.Skip() made drag and drop work but the button would
not be pushed for the life of me. Then I realized that the OnDragEnd
could be hijacked to call the button.Execute. Maybe we could
return FINISH from on_clicked listener instead?

See also https://stackoverflow.com/a/77166531/281545

The MouseCaptureLostEvent is windows only

https://docs.wxpython.org/wx.MouseCaptureLostEvent.html:

> This event is currently emitted under Windows only.

Anyway if self.moved was False it would never release the mouse making
for instance the Settings dialog not responsive. Unconditionally
releasing (not if self.moved only) makes settings dialog (and Bash)
responsive again - probably that's the problem in GTK also?

OnDragStart: not propagating the event makes dnd work and buttons not work

reset dragStart - shouldn't be 0?

Lojack:

Fixup button dragging on windows

- Dragging was working mostly, but reordering buttons was using the
  button under the mouse rather than the dragged button for reordering,
  resulting in a noop
- Had to add back in mouse capture (ugh) to be able to reliably detect
  when the window loses the mouse (ex: alt-tab, popup window, etc) to
  end the dragging state.

Ut: finally this works on all OSes, that's one of the oldest TODOs
and cleans up the way for Lazy - thanks for testing and persevering
everybody!

Co-authored-by: lojack5 <1458329+lojack5@users.noreply.github.com>