WIP Lazy(_AComponent):
I was thinking of this for a while now - there are a few other places
where we could use Lazy:

- gui patcher panels (an opportunity to investigate why they are created
once)
- Link - yep wrapping wx.Menu() is in the last 10% of de-wx'ing and
might make sense to cache those (and clearly differentiate between
stateless and stateful ones)
- Imagewrapper ;)

Currently it's over-fitted to its single use -> StatusBar_Button, but
there is fit quite nicely (note a few methods/class variables gone
and the rest I believe reads more self explanatory than before). For now
I disallow creating the widget through a _native_widget property
access but that might not fit the other usecases (or indeed all
usecases that allow_create is True - note that _append in links
is an allow_create check)

Make _is_created private and instead override the tooltip to not
blow if the _native_widget was not created. See dicussion in RRR

_present -> allow_create and remove useless SetState:

Refactor OBSE handling - remove obseButtons and have the instances check
their _obseTip and the state of the OBSE button to set the correct
tooltip on creation/flip state of obseButton. This will simply
reset the tooltips of all *created* items, probably just resetting some
tooltips to the same str, but the simplification is more than worth it.

Execute -> sb_click

Use the new ring to kill a dragon :P

Use the link.uid in Hide/Unhide that gets us rid of some more dragons

Under #190, #163 #174

We still must make sure the _tip is not empty for the SB settings dialog
or rather that the sb_button_tip returns an unique value