Starfield: Add support for WS version too
Forgot about that one, whoops. No way to test if this is correct,
guess we'll see if bug reports come rolling in.

requirements.txt:
Bump dependencies

pygit2 is only a single fix, while PyInstaller is a big release with
lots of new features and breaking changes. I don't *think* anything
there should affect us, but testing is of course needed.