Fix some _active_entries_to_remove
For Starfield, we need to not remove BlueprintShips-Starfield.esm from
plugins.txt because while it *is* hardcoded to be active, it is *not*
hardcoded to load at a specific load order slot. That means it could
move elsewhere in the load order if we don't write it to plugins.txt.

For Skyrim LE, it turns out we do want to write Update.esm into
plugins.txt. While it's extremely unlikely that it will ever be at any
other position (because Dawnguard.esm has it as a master and so the game
engine will force-load it right before Dawnguard.esm), it is possible
(most notably if the user doesn't have any of the DLCs installed). So
put it into plugins.txt to guarantee it doesn't get loaded later.

This is based on testing done by Ortham.

Under #667