Refactor mergeability checking
There's also a long-term plan to maybe reintroduce the Merge Patches
patcher for SSE/FO4, so this was coming anyways.

I decided to store ints into our settings files for this instead of
the initial strings I experimented with because we will store those
hundreds of times and all those strings would bloat the cache size for
no good reason at that point. Downside is that the int is less readable
in the debugger of course, but that's a minor nuisance at best.