Refactor NoMerge/MergePatches handling
Cut down on duplication a bit: this way, if we ever want to reenable the
patch merger for SSE/FO4, all we have to do is add
MergeabilityCheck.MERGE to their mergeability_checks and everything will
work as intended.

Tested to make sure I'm not saying nonsense here: yup, works perfectly!
Simply adding that enum to SSE's set makes WB check for BP mergeability,
readds the Merge Patches patcher, the BP prompts for merging a test
plugin I made and it merges successfully when building the patch. We can
now reenable that patcher whenever we want :)