Warn about case-insensitive Data folders
And provide help for how to make it case-insensitive. Turns out that you
can even make Windows folders case-sensitive since a Windows 10 update!

We may think about adding some support to WB for emulating
case-insensitivity in the future, but for now users will just have to
work around this. ext4 has the proper, FS-level solution, but most other
filesystems do not have anything like this yet (and for ext4, you have
to specifically enable it via -O flag when mkfs'ing or via later
tune2fs from a recovery medium). ciopfs will work for everyone, but
takes a bit more expertise to set up (and will always be less
elegant/slower/etc.).

We may need to do this for the wineprefix too (i.e. the personal and
local appdata folders), keep an eye out.