Avoid shutil.move race when creating parent dirs
Explicitly call makedirs beforehand, otherwise we can get
FileNotFoundErrors when the parent folder(s) aren't being created fast
enough. This probably happens because move relies on the FS itself to do
this (I don't see any code in the stdlib that actually issues a
mkdir/mkdirs for this) and the FS can schedule to do this async in the
background, leading to a race condition where move *assumes* the folder
has already been created and goes to move the file over.

This honestly seems like a bug in Python if that's what's happening.

Refactor _retry a bit

So that we can avoid the partial, which is actually quite expensive.

Fix callable typing

Rest is moving away from the (deprecated) typing.Callable alias for
collections.abc.Callable (why *that* one is the recommended import, I'll
never understand).