Move parser machinery to a PreParser class:
Decouple PreParser from Bash and WryeParser internals - turns out very
few of the methods of WryeParser did access the instance attributes,
which basically came from the installer. Now these are defined in
PreParser in the abstract and overridden in WryeParser. I wish it
was that way all the countless times I tried to figure out how
BAIN Installer attributes are used in belt - better late than never
however :)
Note that I did move the Begin method up to PreParser although it's
WryeParser territory. However we define a bunch of attributes in there
which made Pycharm complain plus it has no Bash dependencies. There is
much room  for improvement in this essentially py2 code but wizards is
a rabbit hole so I leave the dragon sleep. Apart from the static
`codebox`, that creates an instance of its class that coupled bolt to
belt -> gui - this can be addressed finally.

Mopy/bash/basher/links_init.py: A regression from
2aff9abf4e78e5e0f5bc960f574060f042694e7e - we should really have a
standard fallback for images and with three sizes not one.