Nuke undefinedPaths:
I could not find any occurrence of C:\\Path\\exe.exe
while there are a couple of C:\not\a\valid\path.exe in the ini,
corresponding to where in tooldirs the path is set to undefinedPath.
Since Treed features a path in tooldirs I added that to the ini
(commented  anyway) - not sure why we escape \[ in 'Launch Tree\[d\]'.

Some more nit and fdust from the future (images/app_buttons/tooldirs)

I moved undefinedPath to bolt will come handy (till I remove it) - in
env I used it instead of r'not\a\path' - maybe there are more.