bain: Support installing to multiple destinations
View with whitespace diff off, otherwise you'll get a sea of red and
green with no meaning - I had to insert lots of loops to handle the
multiple destinations, which meant indenting tons of lines.

Sprung up from a seemingly small problem: for some reason, StarUI (a mod
for Starfield) was not getting one of its files installed. The FOMOD GUI
did show it being placed where it was supposed to go, but then it didn't
show up in the Data folder. Turns out it's because the FOMOD specifies
that the file should go into two paths:

 - Interface\StarUI Inventory Presets\StarUI Inventory - Colored.ini
 - Interface\StarUI Inventory (default).ini

The idea being that you get all the presets installed and your chosen
preset also gets set as the default. FOMODs are such a loose format that
this is of course possible. Probably. I've never seen a mod do this
before, but it probably works in Vortex and MO2, which means it's de
facto supported.

BAIN itself doesn't really have problems with handling this, since it
generally maps from destination to source, but our fs operation backends
were not equipped to handle this at all.