fomod: Convert path separators for Linux
On Windows, we never had issues with FOMODs that used '/' as a separator
because normpath replaces slashes with backslashes. But on Linux,
normpath does not convert backslashes. So we have to do this explictly,
otherwise FOMODs may randomly fail to override files from certain
options with files from other options/conditions.