Add typing to event handlers.
This requires a change to event argument processors, they need to return
the arguments as a tuple now, so their types can be (explicitly or inferred)
passed on to the listeners.

Note: this exposes two problems:
1) Lambda functions as argument processors don't have much type information at
all.  See the next commit for a change that can work around this
2) wxPython's typestubs have basically no type information, especially on the
return types, so almost everything gets inferred as "Unknown".