Auto-split the BP
Once the Bashed Patch obtains >255 masters (>253 for ESL games, where
this problem is especially prevalent), there is no way to save the BP
right now. Instead, we show an error message and discard all our work.

That's clearly not ideal, so this commit adds support for auto-splitting
of the BP (at the level of top GRUPs). It also adds some UX improvements
for these split BPs:
 - Clicking 'Rebuild Patch' on any of the parts will rebuild the main
   BP.
 - After rebuilding, unnecessary parts are deleted (with a prompt to let
   the user know and adjust).
 - Renaming the main BP updates the links in the parts.
 - The BP's docs get linked to all the parts as well.

Also fixes List Patch Config not working, which broke back in
2cd690cb86bd0a8d072614d1db9e906880f65a56.

Closes #533