bain: Introduce 'scattered' packages
Adding support for all these weird FOMODs that have their files in
stupid places can't be this simple, right? Apparently it is. Could have
done this a long time ago :P

Needs lots of testing because it *can't* be this easy, surely.

Improve root detection for scattered packages

Of course it isn't that easy :P

Turns out the Unofficial Lux Patches mod has a very stupid structure
where it puts everything inside a root 'fomod' folder. WB detected the
Fomod folder as the root prefix, so could not find the ModuleConfig.xml
anymore (since it was now in effect looking for
fomod/fomod/ModuleConfig.xml).

Also made sure that this works for even *stupider* layouts like

  MyMod
  |- fomod
     |- ModuleConfig.xml
     |- mymod.esp

Because we will probably run into one that is like this.

Closes #670