Hunt down allow_create():
That's something I was wondering about - pass the params in super or set
the img/tip properties after the fact? Turns out the latter is more
correct as it allows us to drop the allow_create checks from the
'version' (called-by sb_tooltip) property. Some extra care must be taken
to make sure wee don't call those on non created buttons. In sb_click we
were performing an extra check just before we launched. This would result
in a nice error message but:

- should be handled by EAFP - difficult to track what error is thrown
in the windows code in _run_exe but anyway it is wrapped in a bare
except and will ShowError - if we figure the windows exceptions we can
Show a particular error for this case and hide the button
- the rest of the code is under the hypothesis that AppButtons are
initialised once. To make these dynamical (in the case an app is missing
on tabbing back in to Bash hide the button for instance) we need to add
some few code - adding it in some places and not others makes reasoning
hard (as if it was easy anyway).

Therefore I nuked the allow_create and made sure we currently don't call
sb_tooltip -> _app_version if the button is not created. This is still
WIP - for one there are buried system calls in there - a property doing
a syscall is a beholder dressed as a mouse.

Now - making the AppLauncher (__future__) an AFile would be the way to
implement refresh but I want us to take the opposite direction. Init the
app_buttons __once__  and let the user modify those from Settings - we can
check if Paths exist when the user launches this Settings page and offer
a refresh_toolbar button in there. For now, removing the allow_create
from sb_click made _app_button_execute redundant - yey, App_Button starts
looking like an API.

Under #570