Very WIP standardize refresh return type:
Don't pay much attention to the API it's poc - refresh return type needs
to be a new type that's for sure. Note the peculiarly named _added,
_updated and _deleted_ finally found their place. They were instance
variables looking for a class.
Note we can't use plain super() in __bool__ overrides here:

File "/Users/ubik-mac13/Dropbox (Personal)/eclipse_workspaces/python/wrye-bash/Mopy/bash/bosh/__init__.py", line 1877, in __bool__
    def __bool__(self): return super().__bool__() or self.ini_changed
                               ^^^^^^^
TypeError: super(type, obj): obj must be an instance or subtype of type

maybe due to `dataclass`? What we really want is to get rid of the
overrides but this will take a bit (for inis, mods are gone in this
merge).

Start of #353 proper.