Work around MO2 VFS bug with webviews
Check is implemented in as `env.is_vfs()`, and triggers
`balt.web_viewer_available()` to be False if MO2 is detected.

Added a fallback in this case for the Bashed Patch result dialog to show
the wtxt instead of the html in this case. The Document Browser could also
use a workaround to show the wtxt instead of html when needed as well, but
not currently implemented.

Infernio:

Some bikeshedding over the name

is_vfs -> in_mo2_vfs, since Linux uses VFSs all the time and none of
those cause a problem. Also move the docstring to windows.py (maybe we
should keep docstrings in both files?)