Properly handle circular masters
I.e. a plugin that depends on itself, either directly (foo.esp has
foo.esp as a master) or transitively (foo.esp has bar.esp as a master,
which has foo.esp as a master - or an even longer chain).

Add a warning to the Plugin Checker, don't allow Change To... to create
this situation and don't allow New Plugin... to create this situation
either.

Utumno:

Mopy/bash/patcher/patchers/mergers.py: the comment was from when I was
trying to load all mods *once* - we would need an override for mergers
but laterz.