_modTimesChange -> unlock_lo - mod redates fix:
Was unlocking in all except one case - in those cases unlocking made
only sense if the game was a timestamp game - I am glad I kept
_modTimesChange and gladder I dropped it

Needs some testing as there are subtle changes in the logic - note that:

mod_links.py: we wouldn't unlock. That was actually a bug (if lock load
order was on) as the redates would silently be reverted in refresh_lo
(with no change shown in the UI) and then a dialog would be displayed on
*next time* we would tab in to Bash (still no change in the UI).
That was probably not the intention :P

Mopy/bash/basher/app_buttons.py: while we did unlock always we did not
forceRefresh - it's a couple stats() added, won't harm, dwarfed by
actually running the manager.

The rest is fine - couple more uses of using_txt_file, one inside the
same (DoSave) scope - good sign.

Mopy/bash/basher/installers_links.py: the (not so) long term goal is to
absorb this into refresh - need to refactor base signature for this.