Drop _reset_info_sets:
It is finally possible to track the info sets so this (very spaghetti)
design (resetting the cache till it's accessed) can be binned. This
makes some calls like new_info more expensive but no worries as the
simplification is worth it definitely for a few ad hoc only calls. Plus
it is not easy to follow the _in_refresh param, which is obviously a
temp hack :P ->

WIP decide what we need to refresh in restore backup:

This is partial (new_info is the hardest to absorb in refresh from the
refresh satellites). No need to further dig in there - refresh will see
an overhaul soon and backup is planned for 320 lol