WIP - Centralize ModInfos refresh logic:
This pile of calls in delete_refresh was probably *the* most ugly wart
from #336 - not least cause most of those were performed anyway in
cached_lo_save_all -> _modinfos_cache_wrapper.
The methods called in _file_or_active_updates have now a single
occurrence - except _update_info_sets which appears in new_info - hmmm,
why not (some of) the rest of the _file_or_active_updates in particular
_refresh_active_no_cp1252? This needs some testing.
Note I added _refreshMergeable (the only camel here, for this reason) in
_modinfos_cache_wrapper. It won't harm as the _lo_cache wrapped calls
are performed explicitly in do once operations so we can afford the
performance but we could be more granular, once it's clear exactly which
changes affect mergeability.